Clarify language re. prioritization of OSS
This commit focuses mainly on the "Selecting Open Source Software First"
section of the "Using Open Source Software" guide. The existing language
described responsibilities of the first four tiers of BIATS
architectural domains. Most of this was removed in favour of focusing on
the titular requirement of selecting OSS first – a responsibility of the
application architecture domain.

The requirement's applicability to both new and existing (e.g.
migration) technology was also added to make the guide as explicit as
the "Requirements" side of the documentation.

Beyond that, this commit does some minor housekeeping by using refs for
repeated links and fixing some typos.