Improvements to identity section (#36)
* Improvements to identity section

I do think that it is useful to be clear that this is beyond just contributing code and documenting commits in a version control system. Users should be encouraged to have their name listed with public accounts. Organizations should be listed too as part of this contribution.

* Update contributing-to-open-source-software.md

Made a few syntax changes to meet our discussion on GCMessage